# openapi.yaml
openapi: 3.0.3
info:
  title: OpenSplit API
  description: Open-source expense-splitting platform API
  version: '1.0'
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.opensplit.example.com
    description: Production

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from /auth/login or /auth/register

  schemas:
    RegisterDto:
      type: object
      required: [email, password, firstName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string

    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    UpdateUserDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        locale:
          type: string
        defaultCurrency:
          type: string
        picture:
          type: string

    CreateGroupDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
        groupType:
          type: string
          enum: [HOME, TRIP, COUPLE, OTHER]
        simplifyByDefault:
          type: boolean
        members:
          type: array
          items:
            type: string

    AddMemberDto:
      type: object
      required: [userId]
      properties:
        userId:
          type: string

    ExpenseShareDto:
      type: object
      required: [userId, paidShare, owedShare]
      properties:
        userId:
          type: string
        paidShare:
          type: number
          minimum: 0
        owedShare:
          type: number
          minimum: 0

    CreateExpenseDto:
      type: object
      required: [description, cost, currencyCode]
      properties:
        groupId:
          type: string
        description:
          type: string
        details:
          type: string
        cost:
          type: number
          minimum: 0
        currencyCode:
          type: string
        categoryId:
          type: integer
        date:
          type: string
          format: date
        repeatInterval:
          type: string
          enum: [NEVER, WEEKLY, FORTNIGHTLY, MONTHLY, YEARLY]
        payment:
          type: boolean
        splitEqually:
          type: boolean
        shares:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseShareDto'

    UpdateExpenseDto:
      type: object
      properties:
        groupId:
          type: string
        description:
          type: string
        details:
          type: string
        cost:
          type: number
          minimum: 0
        currencyCode:
          type: string
        categoryId:
          type: integer
        date:
          type: string
          format: date
        repeatInterval:
          type: string
          enum: [NEVER, WEEKLY, FORTNIGHTLY, MONTHLY, YEARLY]
        payment:
          type: boolean
        splitEqually:
          type: boolean
        shares:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseShareDto'

    CreateFriendDto:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string

    CreateCommentDto:
      type: object
      required: [content]
      properties:
        content:
          type: string

    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
        timestamp:
          type: string
          format: date-time

paths:
  # ─── Auth ───────────────────────────────────────────────
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive API key
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/rotate-key:
    post:
      tags: [Auth]
      summary: Rotate the current API key
      responses:
        '200':
          description: New API key generated
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Users ──────────────────────────────────────────────
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: Current user profile
        '401':
          description: Unauthorized

  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
        '404':
          description: User not found
    patch:
      tags: [Users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: User updated
        '404':
          description: User not found

  # ─── Groups ─────────────────────────────────────────────
  /groups:
    get:
      tags: [Groups]
      summary: List all groups for the current user
      responses:
        '200':
          description: List of groups
    post:
      tags: [Groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupDto'
      responses:
        '201':
          description: Group created

  /groups/{id}:
    get:
      tags: [Groups]
      summary: Get group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group details
        '404':
          description: Group not found
    delete:
      tags: [Groups]
      summary: Soft delete a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group deleted
        '404':
          description: Group not found

  /groups/{id}/restore:
    post:
      tags: [Groups]
      summary: Restore a deleted group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Group restored
        '404':
          description: Group not found

  /groups/{id}/members:
    post:
      tags: [Groups]
      summary: Add a member to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberDto'
      responses:
        '201':
          description: Member added
        '404':
          description: Group not found
        '409':
          description: Member already exists

  /groups/{id}/members/{userId}:
    delete:
      tags: [Groups]
      summary: Remove a member from a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member removed
        '404':
          description: Group or member not found

  # ─── Expenses ───────────────────────────────────────────
  /expenses:
    get:
      tags: [Expenses]
      summary: List expenses
      parameters:
        - name: group_id
          in: query
          schema:
            type: string
        - name: friend_id
          in: query
          schema:
            type: string
        - name: dated_after
          in: query
          schema:
            type: string
            format: date
        - name: dated_before
          in: query
          schema:
            type: string
            format: date
        - name: updated_after
          in: query
          schema:
            type: string
            format: date
        - name: updated_before
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: List of expenses
    post:
      tags: [Expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseDto'
      responses:
        '201':
          description: Expense created

  /expenses/{id}:
    get:
      tags: [Expenses]
      summary: Get expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expense details
        '404':
          description: Expense not found
    patch:
      tags: [Expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpenseDto'
      responses:
        '200':
          description: Expense updated
        '404':
          description: Expense not found
    delete:
      tags: [Expenses]
      summary: Soft delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expense deleted
        '404':
          description: Expense not found

  /expenses/{id}/restore:
    post:
      tags: [Expenses]
      summary: Restore a deleted expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expense restored
        '404':
          description: Expense not found

  # ─── Comments ───────────────────────────────────────────
  /expenses/{expenseId}/comments:
    get:
      tags: [Comments]
      summary: List comments for an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments
        '404':
          description: Expense not found
    post:
      tags: [Comments]
      summary: Add a comment to an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: Comment created
        '404':
          description: Expense not found

  /comments/{id}:
    delete:
      tags: [Comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comment deleted
        '404':
          description: Comment not found

  # ─── Friends ────────────────────────────────────────────
  /friends:
    get:
      tags: [Friends]
      summary: List all friends
      responses:
        '200':
          description: List of friends
    post:
      tags: [Friends]
      summary: Add a friend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFriendDto'
      responses:
        '201':
          description: Friend added
        '409':
          description: Already friends

  /friends/{id}:
    get:
      tags: [Friends]
      summary: Get friend by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Friend details
        '404':
          description: Friend not found
    delete:
      tags: [Friends]
      summary: Remove a friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Friend removed
        '404':
          description: Friend not found

  # ─── Notifications ──────────────────────────────────────
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      parameters:
        - name: updated_after
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: List of notifications

  # ─── Currencies ─────────────────────────────────────────
  /currencies:
    get:
      tags: [Currencies]
      summary: List all currencies
      responses:
        '200':
          description: List of currencies

  # ─── Categories ─────────────────────────────────────────
  /categories:
    get:
      tags: [Categories]
      summary: List all categories
      responses:
        '200':
          description: List of categories